Migrate to Angular standalone
