+++ packages/desktop/src/renderer/app/components/menus/callbacks-menu/callbacks-menu.component.ts
@@ -1,9 +1,18 @@
+import {
+  AsyncPipe,
+  LowerCasePipe,
+  NgClass,
+  NgFor,
+  NgIf,
+  UpperCasePipe
+} from '@angular/common';
 import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import {
   Callback,
   ReorderAction,
   ReorderableContainers
 } from '@mockoon/commons';
+import { NgbTooltip } from '@ng-bootstrap/ng-bootstrap';
 import { Observable } from 'rxjs';
 import {
   combineLatestWith,
@@ -11,7 +20,16 @@ import {
   filter,
   map
 } from 'rxjs/operators';
-import { DropdownMenuItem } from 'src/renderer/app/components/dropdown-menu/dropdown-menu.component';
+import {
+  DropdownMenuComponent,
+  DropdownMenuItem
+} from 'src/renderer/app/components/dropdown-menu/dropdown-menu.component';
+import { FilterComponent } from 'src/renderer/app/components/filter/filter.component';
+import { SvgComponent } from 'src/renderer/app/components/svg/svg.component';
+import { DraggableDirective } from 'src/renderer/app/directives/draggable.directive';
+import { DropzoneDirective } from 'src/renderer/app/directives/dropzone.directive';
+import { ResizeColumnDirective } from 'src/renderer/app/directives/resize-column.directive';
+import { ScrollWhenActiveDirective } from 'src/renderer/app/directives/scroll-to-active.directive';
 import { FocusableInputs } from 'src/renderer/app/enums/ui.enum';
 import { textFilter, trackByUuid } from 'src/renderer/app/libs/utils.lib';
 import { EnvironmentsService } from 'src/renderer/app/services/environments.service';
@@ -26,7 +44,22 @@ type dropdownMenuPayload = { callbackUuid: string };
   templateUrl: './callbacks-menu.component.html',
   styleUrls: ['./callbacks-menu.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
-  standalone: false
+  imports: [
+    NgIf,
+    NgbTooltip,
+    SvgComponent,
+    FilterComponent,
+    NgFor,
+    DraggableDirective,
+    DropzoneDirective,
+    ScrollWhenActiveDirective,
+    NgClass,
+    DropdownMenuComponent,
+    ResizeColumnDirective,
+    AsyncPipe,
+    UpperCasePipe,
+    LowerCasePipe
+  ]
 })
 export class CallbacksMenuComponent implements OnInit {
   public settings$: Observable<Settings>;
