+++ packages/desktop/src/renderer/app/components/modals/confirm-modal/confirm-modal.component.ts
@@ -1,3 +1,4 @@
+import { AsyncPipe, NgFor, NgIf } from '@angular/common';
 import {
   ChangeDetectionStrategy,
   Component,
@@ -6,6 +7,7 @@ import {
 } from '@angular/core';
 import { Observable, Subject } from 'rxjs';
 import { takeUntil, tap } from 'rxjs/operators';
+import { SvgComponent } from 'src/renderer/app/components/svg/svg.component';
 import { ConfirmModalPayload } from 'src/renderer/app/models/ui.model';
 import { UIService } from 'src/renderer/app/services/ui.service';
 
@@ -13,7 +15,7 @@ import { UIService } from 'src/renderer/app/services/ui.service';
   selector: 'app-confirm-modal',
   templateUrl: './confirm-modal.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
-  standalone: false
+  imports: [NgIf, NgFor, SvgComponent, AsyncPipe]
 })
 export class ConfirmModalComponent implements OnInit, OnDestroy {
   public confirmModalPayload$: Observable<ConfirmModalPayload>;
