+++ packages/desktop/src/renderer/app/directives/scroll-to-active.directive.ts
@@ -4,10 +4,7 @@ import { Subject } from 'rxjs';
 /**
  * This directive is used to scroll the element into view when it becomes active (class added) or when the element is created with the active class (element insertion, click or class change from outside, like the command palette).
  */
-@Directive({
-  selector: '[appScrollWhenActive]',
-  standalone: false
-})
+@Directive({ selector: '[appScrollWhenActive]' })
 export class ScrollWhenActiveDirective implements AfterViewInit, OnDestroy {
   private destroy$ = new Subject<void>();
   private mutationObserver = new MutationObserver(
